requirements.txt
Flask==2.1.2
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
python-dotenv==0.20.0
pusher==5.0.0